% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anyWhich.R
\name{anyGeq}
\alias{anyGeq}
\title{Any without logical creation}
\usage{
anyGeq(x, a)
}
\arguments{
\item{x}{An integer vector.}

\item{a}{An integer.}
}
\value{
0 if none true or the index of the first match.
}
\description{
Any without logical creation
}
