% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AnyWhich}
\alias{AnyWhich}
\title{Quickly verify (and locate) the existence of a breach.}
\arguments{
\item{x}{An integer vector.}

\item{a}{A (single) integer. That which is to be compared.}

\item{gt, lt, eq}{Booleans, whether or not the comparison is greater than, less than, or equal to.
Only \code{gt} and \code{lt} are mutually exclusive.}
}
\description{
Used when a single instance is likely to occur and be important to detect quickly
(in a sufficiently large integer vector).
}
