% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awote.R
\name{awote}
\alias{awote}
\title{AWOTE}
\usage{
awote(Date = NULL, fy.year = NULL, Sex = NULL, isAdult = TRUE,
  rollDate = "nearest", isOrdinary = TRUE)
}
\arguments{
\item{Date, fy.year}{When the AWOTE is desired.}

\item{Sex}{(If \code{NULL} )}
}
\description{
Adult weekly ordinary-time earnings
}
