% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bto.R
\name{bto}
\alias{bto}
\title{Beneficiary tax offset}
\usage{
bto(benefit_amount, fy.year)
}
\arguments{
\item{benefit_amount}{The amount of Tax Offsetable benefit received by the taxpayer during the income year.}

\item{fy.year}{The income year.}
}
\value{
The beneficiary tax offset.
}
\description{
Beneficiary tax offset
}
\section{WARNING}{

This function disagrees with the ATO online calculator.
}

