% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carers_allowance.R
\name{carers_allowance}
\alias{carers_allowance}
\title{Carers allowance}
\usage{
carers_allowance(Date = NULL, fy.year = NULL, per = c("year",
  "fortnight"))
}
\arguments{
\item{Date, fy.year}{The timing of the allowance.}

\item{per}{Frequency of the payment.}
}
\value{
The carer's payment, if eligible.
}
\description{
Carers allowance
}
