% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carers_payment.R
\name{carers_payment}
\alias{carers_payment}
\title{Carers payment}
\usage{
carers_payment(Date = NULL, fy.year = NULL)
}
\arguments{
\item{Date, fy.year}{The timing of the payment.}
}
\value{
The carer's payment, if eligible.
}
\description{
Carers payment
}
