% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_inflator.R
\name{CG_population_inflator}
\alias{CG_inflator}
\alias{CG_population_inflator}
\title{Forecasting capital gains}
\usage{
CG_population_inflator(x = 1, from_fy, to_fy, estimator = "mean",
  prediction_interval = 80)

CG_inflator(x = 1, from_fy, to_fy, ...)
}
\arguments{
\item{x}{To be inflated.}

\item{from_fy, to_fy}{Financial years designating the inflation period.}

\item{estimator}{One of \code{"mean"}, \code{"lower"}, \code{"upper"}. What estimator to use in forecasts. \code{"lower"} and \code{"upper"} give the lower and upper boundaries of the prediction interval,}

\item{prediction_interval}{A single value specifying the prediction interval to be used if \code{estimator} is \code{"lower"} or \code{"upper"}. By default, \code{80}, i.e. an 80\% prediction interval.}

\item{...}{Passed to \code{CG_population_inflator}}
}
\value{
An estimate of \code{x} inflated to \code{to_fy}
}
\description{
Forecasting capital gains
}

