% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgt_inflator.R
\name{CGT_inflator}
\alias{CGT_inflator}
\title{CGT inflator}
\usage{
CGT_inflator(cgt = 1, from_fy, to_fy)
}
\arguments{
\item{cgt}{The amount to be inflated.}

\item{from_fy}{The financial year of the current value of CGT.}

\item{to_fy}{The financial year for which CGT is intended to be inflated.}
}
\value{
An estimate of the CGT payable in \code{to_fy}, inflated by relative tax expenditures.
}
\description{
For determing the size of capital gains tax in future years from previous values (i.e. from sample files)
}

