% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child_care_subsidy.R
\name{child_care_subsidy}
\alias{child_care_subsidy}
\title{Child care Subsidy}
\usage{
child_care_subsidy(family_annual_income, child_age, type_of_day_care)
}
\arguments{
\item{family_annual_income}{}
}
\value{
jdjd
}
\description{
Child care Subsidy
}
