% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean_bibtex_output_from_endnote.R
\name{clean_bibtex_output_from_endnote}
\alias{clean_bibtex_output_from_endnote}
\title{endnote_to_bibtex}
\usage{
clean_bibtex_output_from_endnote(filename,
  outname = paste0(gsub("(.*)\\\\.txt$", "\\\\1", filename), "cleaned.txt"))
}
\arguments{
\item{filename}{the exported txt filename carried from Endnote via BibTeX Export utility that is dirty and must be cleaned}

\item{outname}{the clean(er) text file that is carried out by this process}
}
\description{
endnote_to_bibtex
}

