% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_general_date.R
\name{cpi_general_date}
\alias{cpi_general_date}
\title{CPI for general dates}
\usage{
cpi_general_date(nominal.price, nominal.date, target.date, ...)
}
\arguments{
\item{nominal.price}{(numeric) the nominal prices to be converted to a real price}

\item{nominal.date}{(character, date-like) the 'date' contemporaneous to \code{nominal price}. The acceptable forms are 'YYYY', 'YYYY-YY' (financial year), 'YYYY-MM-DD'.}

\item{target.date}{(character, date-like) the date at which the real price is valued (where the nominal price equals the real price). Same forms as for \code{nominal.date}}

\item{...}{other arguments passed to \code{cpi_inflator_quarters}}
}
\value{
a vector of real prices in target.date dollars.
}
\description{
CPI for general dates
}

