% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_inflator_quarters.R
\name{cpi_inflator_quarters}
\alias{cpi_inflator_quarters}
\title{CPI inflator when dates are nice}
\usage{
cpi_inflator_quarters(nominal.price, nominal.date, target.date,
  adjustment = "none")
}
\arguments{
\item{nominal.price}{(numeric) the nominal prices to be converted to a real price}

\item{nominal.date}{(date in quarters) the dates contemporaneous to the prices in nominal.price. Must be of the form "YYYY-Qq" e.g. "1066-Q2"}

\item{target.date}{(date in quarters) the date to be inflated to, where nominal price = real price. Must be of the form "YYYY-Qq" e.g. "1066-Q2"}

\item{adjustment}{Should there be an adjustment made to the index? Adjustments include 'none' (no adjustment), 'seasonal', or 'trimmed mean'.}
}
\value{
a vector of real prices
}
\description{
CPI inflator when dates are nice
}
\details{
this uses the ABS stat SDMX interface, which will form the bulk of the execution time.
}

