% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tax_benefit.R
\name{family_tax_benefit}
\alias{family_tax_benefit}
\title{FAMILY TAX BENEFIT}
\usage{
family_tax_benefit(child_age = 1, child_in_secondary_school = FALSE,
  child_in_care = FALSE)
}
\description{
FAMILY TAX BENEFIT
}
