% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreign_income_tax_offset.R
\name{foreign_income_tax_offset}
\alias{foreign_income_tax_offset}
\title{Foreign Income Tax Offset}
\usage{
foreign_income_tax_offset(tax_payable, fy_year, foreign_assessable_income,
  other_foreign_income, foreign_entitlements = 0)
}
\arguments{
\item{tax_payable}{The income tax payable (including Medicare levy, Medicare levy surcharge and temporary budget repair levy) for the relevant income year, excluding penalties and interest and disregarding any tax offsets.}

\item{foreign_assessable_income}{The amount of taxable income on which foreign income has been paid.}

\item{other_foreign_income}{Other income or gains from a non-Australian source.}

\item{foreign_entitlements}{The sum of 
\itemize{
\item{debt deductions attributable to your overseas permanent establishment}
\item{any other deductions (other than debt deductions) that are reasonably related to any amount covered by the first dot point above}
\item{an amount of the foreign loss component of one or more tax losses deducted in the income year.}
}}
}
\description{
Foreign Income Tax Offset
}
