% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gni_qtr.R
\name{gni_qtr}
\alias{gni_qtr}
\title{Gross National Income, Australia}
\usage{
gni_qtr(date, roll = "nearest")
}
\arguments{
\item{date}{A Date vector or character coercible thereto.}

\item{roll}{Passed to \code{data.table} when joining.}
}
\value{
For \code{gni_qtr}, the quarterly GNI for the nearest quarter date.
}
\description{
Gross national income, at contemporaneous prices (called 'current prices' by the ABS).
}

