% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpalx.R
\name{gpalx}
\alias{gpalx}
\title{The extended colour palette}
\usage{
gpalx(n)
}
\arguments{
\item{n}{a positive integer}
}
\value{
A vector of n HTML colours.
}
\description{
The extended colour palette
}
\details{
This function interpolates the original Grattan palette for a large (possibly infinite) number of factors.
}

