% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gratplot_discrete.R
\name{gratplot_discrete}
\alias{gratplot_discrete}
\title{attempt to harness ggplot but set as much to default as possible}
\usage{
gratplot_discrete(...)
}
\description{
attempt to harness ggplot but set as much to default as possible
}

