% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan-package.R
\docType{package}
\name{grattan-package}
\alias{grattan-package}
\alias{grattan}
\title{The grattan package.}
\description{
Grattan package
}
\details{
Tax modelling and other common tasks for Australian policy analysts, 
in support of the Grattan Institute, Melbourne. 
<https://grattan.edu.au>
}
\section{Package options}{

\describe{
\item{\code{grattan.verbose}}{(\code{FALSE}) Emit diagnostic messages (via \code{cat()}))}
\item{\code{grattan.assume1901_2100}}{(\code{TRUE}) Assume \code{yr2fy} receives an integer >= 1901 and <= 2100.}
\item{\code{grattan.taxstats.lib}}{Package library into which \code{taxstats} packages will
be installed. If \code{NULL}, a temporary directory is used.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HughParsonage/grattan}
  \item \url{https://hughparsonage.github.io/grattan/}
  \item Report bugs at \url{https://github.com/HughParsonage/grattan/issues}
}

}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}

\email{hugh.parsonage@gmail.com}
}
\keyword{package}
