% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grattan_waterfall.R
\name{grattan waterfall}
\alias{grattan waterfall}
\alias{grattan_waterfall}
\title{Create waterfall charts}
\usage{
grattan_waterfall(.data = NULL, values, labels, calc_total = FALSE,
  total_rect_color = gpal(6)[1], total_rect_text_color = "white",
  fill_colours = gpalx(length(values)), fill_by_sign = TRUE, dark = FALSE,
  rect_width = 0.7, rect_border = "black", draw_lines = TRUE,
  lines_anchors = c("centre", "centre"), linetype = "dashed",
  ggplot_object_name = "p")
}
\arguments{
\item{.data}{a placeholder for piped operations; cannot be used}

\item{values}{a numeric vector making up the heights of the rectangles in the waterfall}

\item{labels}{the labels corresponding to each vector, marked on the x-axis}

\item{calc_total}{(logical) should the final pool of the waterfall be calculated (and placed on the chart)}

\item{total_rect_color}{the color of the final rectangle}

\item{total_rect_text_color}{the color of the final rectangle's label text}

\item{fill_by_sign}{(logical) should positive and negative values each have the same colour?}

\item{dark}{(logical) (only for fill_by_sign) should the dark palette be used?}

\item{rect_width}{(numeric) the width of the rectangle, relative to the space between each label factor}

\item{rect_border}{the border around each rectangle. Choose NA if no border is desired.}

\item{draw_lines}{(logical) should lines be drawn between successive rectangles}

\item{lines_anchors}{a character vector of length two specifying the horizontal placement of the drawn lines relative to the preceding and successive rectangles, respectively}

\item{linetype}{the linetype for the draw_lines}

\item{ggplot_object_name}{(character) a quoted valid object name to which ggplot layers may be addded after the function has run}
}
\description{
Create waterfall charts
}

