% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_frac.R
\name{Grattan_frac}
\alias{Grattan_frac}
\alias{grattan_frac}
\title{English prose for small numbers}
\usage{
Grattan_frac(number, hedges = c("Almost", "Over"))

grattan_frac(number, hedges = c("Almost", "Over"))
}
\arguments{
\item{number}{A length-one numeric vector.}

\item{hedges}{A length-2 character vector giving the words for a hedge (see \url{https://en.wikipedia.org/wiki/Hedge_(linguistics)}), 
included not to mislead but to sound natural without being inaccurate. 
If \code{NULL}, the nearest fraction regardless is used.}
}
\value{
For \code{Grattan_frac} and \code{grattan_frac}, the nearest fraction as an English word, sentence-case or lower-case respectively. 
If \code{hedges} is supplied, a phrase including the next fraction above or below it as the phrase dictates.
}
\description{
When a small number is requested inline, it is often preferable to use an English expression rather than the numeral.
Further, referential purity can be lost when numbers are transformed to suit printing (say by multiplying by 100 to use 'per cent').
A complement to \code{\link{texNum}}. 

These functions are all of the form \code{grattan_*} to avoid namespace clashes.
}

