% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_frac_percent.R
\name{grattan_percent}
\alias{grattan_percent}
\title{English prose for small numbers}
\usage{
grattan_percent(number, digits = 1, .percent.suffix = "~per~cent")
}
\arguments{
\item{number}{A length-one numeric vector.}

\item{digits}{Passed to \code{scales::comma}.}

\item{.percent.suffix}{What to use for the suffix of the number.}

\item{hedges}{A length-2 character vector giving the words for a hedge (see \url{https://en.wikipedia.org/wiki/Hedge_(linguistics)}), 
included not to mislead but to sound natural without being inaccurate. 
If \code{NULL}, the nearest fraction regardless is used.}
}
\value{
For \code{grattan_percent}, the number multiplied by 100, with rounding provided by \code{scales::comma} and then the suffix. 

For \code{Grattan_frac} and \code{grattan_frac}, the nearest fraction as an English word, sentence-case or lower-case respectively. 
If \code{hedges} is supplied, a phrase including the next fraction above or below it as the phrase dictates.
}
\description{
When a small number is requested inline, it is often preferable to use an English expression rather than the numeral.
Further, referential purity can be lost when numbers are transformed to suit printing (say by multiplying by 100 to use 'per cent').
A complement to \code{\link{texNum}}. 

These functions are all of the form \code{grattan_*} to avoid namespace clashes.
}

