% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/grey.pal.R
\docType{data}
\name{Grey palette}
\alias{Grey palette}
\alias{grey.pal}
\title{The grey palette}
\format{\preformatted{ chr [1:6] "#D8D8D8" "#ADADAD" "#818181" "#575757" "#2B2B2B" ...
}}
\usage{
grey.pal
}
\description{
The grey palette
}
\keyword{datasets}

