% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grplot.R
\name{grplot}
\alias{grplot}
\title{Attempt to harness ggplot but set as much to default as possible}
\usage{
grplot(..., reverse = FALSE)
}
\arguments{
\item{...}{arguments passed to ggplot}

\item{reverse}{(logical) Option to reverse the default palette.}
}
\description{
Attempt to harness ggplot but set as much to default as possible
}

