% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{tax function}
\usage{
income_tax(income, fy.year, age = 42, family_status = "individual",
  n_dependants = 0L, sample_file, .dots.ATO = NULL,
  return.mode = "numeric", allow.forecasts = FALSE)
}
\arguments{
\item{income}{the personal assessable income}

\item{fy.year}{the financial year in which the income was earned}

\item{age}{the individual's age}

\item{family_status}{For medicare and sapto purposes. Still in development.}

\item{n_dependants}{An integer for the number of children of the taxpayer (for the purposes of the Medicare levy).}

\item{sample_file}{(Not yet used) A sample file \code{data.table} for which the income tax payable is desired on each row.}

\item{.dots.ATO}{A data.frame that contains additional information about the individual's circumstances, with columns the same as in the ATO sample files.}

\item{return.mode}{use numeric (integer not yet implemented).}

\item{allow.forecasts}{should dates beyond 2014-15 be permitted?}
}
\value{
the total personal income tax payable
}
\description{
tax function
}
\details{
The function 'rolling' is inflexible by design. It is designed to guarantee the correct tax payable in a year.
}
\author{
Tim Cameron, Brendan Coates, Hugh Parsonage, William Young
}

