% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IncomeTax}
\alias{IncomeTax}
\title{IncomeTax}
\arguments{
\item{x}{Taxable income.}

\item{thresholds}{Lower brackets of the tax tables.}

\item{rates}{Marginal rates}
}
\description{
Calculates the ordinary tax payable given income and tax thresholds and rates.
Basic, designed for performance.
}
