% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflator.R
\name{inflator}
\alias{inflator}
\title{Inflate using a general index}
\usage{
inflator(x = 1, from, to, inflator_table)
}
\arguments{
\item{x}{The vector to be inflated.}

\item{from}{The contemporaneous time of x.}

\item{to}{The target time to which x is to be inflated.}

\item{inflator_table}{A \code{data.table} having columns \code{Index} and \code{Time}.}
}
\value{
A vector of inflated values.
}
\description{
Inflate using a general index
}

