% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_taxstats.R
\name{install_taxstats}
\alias{install_taxstats}
\title{Install 'taxstats' files}
\usage{
install_taxstats(pkg = c("taxstats"), ...)
}
\arguments{
\item{pkg}{The package to install such as \code{"taxstats"} or \code{"taxstats1516"}.}

\item{...}{Arguments passed to \code{\link[utils]{install.packages}}.}
}
\description{
The taxstats packages provide the sample files as released by the ATO. These packages are used for testing, but are not available through CRAN as they are too large.
}
