% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inverse_income.R
\name{inverse_income}
\alias{inverse_income}
\title{Inverse income tax function}
\usage{
inverse_income(tax, fy.year = "2012-13")
}
\arguments{
\item{tax}{the tax payable}
}
\value{
the taxable income (integer) given the tax payable for the financial year, acurrate to the nearest integer
}
\description{
Inverse income tax function
}

