% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fy.year.R
\name{is.fy}
\alias{is.fy}
\title{Convenience functions for creating FY}
\usage{
is.fy(fy.yr)
}
\arguments{
\item{fy.yr}{A string suspected to be a financial year.}

\item{yr_ending}{An integer representing a year.}

\item{date}{A string or date for which the financial year is desired.}
}
\description{
Convenience functions for creating FY
}

