% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_inflator.R
\name{lf_inflator}
\alias{lf_inflator}
\alias{lf_inflator_fy}
\title{Labour force inflators}
\usage{
lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy,
  useABSConnection = FALSE, allow.projection = TRUE, use.month = "01")

lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date)
}
\arguments{
\item{labour_force}{a numeric vector}

\item{from_fy}{Financial year of \code{labour_force}.}

\item{to_fy}{Financial year for which the labour force is predicted.}

\item{useABSConnection}{Should an sdmx connection be used to get ABS data?}

\item{allow.projection}{Logical. Should projections be allowed?}

\item{use.month}{A 2-character string for the month in which the labour force should be used.}

\item{from_date}{The date of \code{labour_force}.}

\item{to_date}{dates as a character vector}
}
\value{
the relative labour force between to_date and for_date, multiplied by labour_force.
}
\description{
Labour force inflators
}
\examples{
lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy = "2013-14")
\dontrun{
lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date = "2014-06-30")
}
}

