% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LITO.R
\name{lito}
\alias{lito}
\alias{lmito}
\title{Low Income Tax Offset}
\usage{
lito(income, fy.year = NULL)

lmito(income, fy.year = NULL)
}
\arguments{
\item{income}{The income on which the offset is applied.}

\item{fy.year}{The financial year for which the LITO is desired.}
}
\value{
The LITO or LMITO for the given income and tax year.
}
\description{
The Low Income Tax Offset (LITO) is a non-refundable tax offset to reduce ordinary personal income tax for low-income earners.

N.B. Since v2.0.0, \code{lito} only calculates the actual LITO, rather than
an offset with custom parameters. For such functionality, use (unexported) \code{Offset}.
}
