% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_super_contr_base}
\alias{max_super_contr_base}
\title{Maximum superannuation contribution base}
\format{
A data frame with 25 rows and 2 variables:
\describe{
 \item{fy_year}{The financial year.}
 \item{max_sg_per_qtr}{Maximum superannuation guarantee per quarter.}
}
}
\source{
ATO.
}
\usage{
max_super_contr_base
}
\description{
Data maximum super contribution base.
}
\keyword{datasets}
