% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_rent_assistance.R
\name{model_rent_assistance}
\alias{model_rent_assistance}
\title{Model Rent Assistance}
\usage{
model_rent_assistance(sample_file, .Prop_rent_paid_by_RA, Max_rate = NULL,
  Min_rent = NULL)
}
\description{
Model Rent Assistance
}
\examples{

}
