% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MultiOffset}
\alias{MultiOffset}
\title{Multioffset}
\arguments{
\item{x}{A numeric vector that the thresholds refer to.}

\item{first_offset}{\code{numeric(1)} The value of the first offset,
the value of the offset for \code{x < first(thresholds)}.}

\item{thresholds}{The thresholds for changes in the offset.}

\item{tapers}{The tapers that apply \strong{above} the corresponding element
of \code{thresholds}.}

\item{above_zero}{(bool) Should negative offsets be set to zero?}
}
\description{
An offset with multiple thresholds and tapers.
}
\examples{
MultiOffset(c(36e3, 37e3, 38e3, 47e3, 48e3, 49e3), 
            255,
            c(37e3, 48e3, 90e3),
            c(0.075, 0, -0.03))
}
