% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_medicare_levy.R
\name{new_medicare_levy}
\alias{new_medicare_levy}
\title{New medicare levy}
\usage{
new_medicare_levy(parameter_table, elderly_age_is = 65)
}
\arguments{
\item{parameter_table}{A \code{data.table} containing 
\describe{
\item{\code{elderly}}{A logical column whether to designate a different medicare.}
\item{\code{family_status}}{'family' for a different threshold for dependants}
\item{\code{lower_threshold}}{What is the lower medicare threshold, below which no medicare levy is applied, above which a tapering rate applies.}
\item{\code{taper}}{What is the taper above \code{lower_threshold}.}
\item{\code{rate}}{The medicare levy applicable above the medicare thresholds.}
\item{\code{lower_up_for_each_child}}{How much the lower threshold should increase with each \code{n_dependants}.}
}}

\item{elderly_age_is}{What age does the elderly row apply?}
}
\value{
A function similar to \code{medicare_levy}.
}
\description{
Use a different way to calculate medicare levy.
}

