% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sapto.R
\name{new_sapto}
\alias{new_sapto}
\title{New SAPTO thresholds}
\usage{
new_sapto(rebate_income, new_sapto_tbl, sapto.eligible = TRUE, fill = 0,
  family_status = "single")
}
\arguments{
\item{rebate_income}{The rebate income of the individual.}

\item{new_sapto_tbl}{Having the same columns as \code{grattan:::sapto_tbl}, keyed on family_status.}

\item{sapto.eligible}{Is the individual eligible for sapto?}

\item{fill}{If SAPTO was not applicable, what value should be used?}

\item{family_status}{Family status of the individual.}
}
\description{
New SAPTO thresholds
}
