% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstart_allowance.R
\name{newstart_allowance}
\alias{newstart_allowance}
\title{Newstart allowance}
\source{
\url{http://classic.austlii.edu.au/au/legis/cth/consol_act/ssa1991186/s1068.html}
}
\usage{
newstart_allowance(ordinary_income, partner_income = 0,
  partner_eligible = FALSE, age = NULL, Date = NULL,
  income_free_area = NULL, income_threshold = NULL, Rate = NULL,
  taper1 = 0.5, taper2 = 0.6, per = "fortnight")
}
\arguments{
\item{ordinary_income}{'Ordinary income' within the meaning of s. 1068-G1 of the \emph{Social Security Act 1991}.}

\item{partner_income}{The partner's income.}

\item{age}{The individual's age. Different rates apply to those aged 18 to 20.}

\item{per}{Fortnight, the default, specified the units the inputs and outputs are returned.}
}
\description{
Newstart allowance
}
