% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstart_income.R
\name{newstart_income}
\alias{newstart_income}
\title{Newstart income}
\usage{
newstart_income(fortnightly.income)
}
\arguments{
\item{fortnightly.income}{income from other sources}
}
\value{
the amount the Newstart recipient would receive in total
}
\description{
Newstart income
}
\note{
doesn't include Rent Assistance
}

