% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialfunctions.R
\name{npv}
\alias{npv}
\title{Financial functions}
\usage{
npv(rate, values)
}
\description{
\url{http://cvs.moodle.org/contrib/patches/question_calculated_extended/calculated/packages/financial/financial_class.php?view=co}
}
\author{
Enrique Garcia M. \email{egarcia@egm.as}

Karsten W. \email{k.weinert@gmx.net}
}

