% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Offset}
\alias{Offset}
\title{General offset in C++}
\arguments{
\item{x}{A vector of incomes etc.}

\item{y}{The maximum offset available; the offset when \code{x} is zero.}

\item{a}{The maximum value of \code{x} at which the maximum offset is available.}

\item{m}{The taper rate (the \strong{negative} slope).}
}
\description{
Calculate the offset given a threshold, a maximum offset, and a taper.
}
