% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmax3}
\alias{pmax3}
\title{Threeway parallel maximum}
\arguments{
\item{x, y, z}{Numeric vectors of identical lengths.}
}
\value{
The parallel maximum of the vectors.
}
\description{
Returns the parallel maximum of three
}
