% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmaxC.R
\name{pmaxC}
\alias{pmaxC}
\alias{pmax3}
\alias{pminC}
\alias{pmaxV}
\alias{pminV}
\title{Objects imported from other packages}
\usage{
pmaxC(
  x,
  a,
  in_place = FALSE,
  keep_nas = FALSE,
  dbl_ok = NA,
  nThread = getOption("hutilscpp.nThread", 1L)
)

pmax3(x, y, z, in_place = FALSE)

pminC(
  x,
  a,
  in_place = FALSE,
  keep_nas = FALSE,
  dbl_ok = NA,
  nThread = getOption("hutilscpp.nThread", 1L)
)

pmaxV(
  x,
  y,
  in_place = FALSE,
  dbl_ok = TRUE,
  nThread = getOption("hutilscpp.nThread", 1L)
)

pminV(
  x,
  y,
  in_place = FALSE,
  dbl_ok = TRUE,
  nThread = getOption("hutilscpp.nThread", 1L)
)
}
\description{
Objects imported from other packages
}
