% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pminC}
\alias{pminC}
\title{Parallel maximum}
\arguments{
\item{x}{A numeric vector.}

\item{a}{A single numeric value.}
}
\value{
The parallel minimum of the input values. The \code{0} versions are shortcuts for \code{a = 0}.
}
\description{
A faster \code{pmin()}.
}
\note{
This function will always be faster than \code{pmin(x, a)} when \code{a} is a single value, but can be slower than \code{pmin.int(x, a)} when \code{x} is short. Use this function when comparing a numeric vector with a single value.
}
