% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pminV}
\alias{pminV}
\title{Parallel maximum}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}
}
\value{
The parallel maximum of the input values.
}
\description{
A faster \code{pmin()}.
}
