% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prohibit_length0_vectors.R
\name{prohibit_length0_vectors}
\alias{prohibit_length0_vectors}
\title{Prohibit zero lengths}
\usage{
prohibit_length0_vectors(...)
}
\arguments{
\item{...}{A list of vectors}
}
\value{
An error message if any of vector has zero length.
}
\description{
Tests whether any vectors have zero length.
}

