% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prohibit_unequal_length_vectors.R
\name{prohibit_unequal_length_vectors}
\alias{prohibit_unequal_length_vectors}
\title{Prohibit unequal length vectors}
\usage{
prohibit_unequal_length_vectors(...)
}
\arguments{
\item{...}{Vectors to test.}
}
\value{
An error message unless all of \code{...} have the same length in which case \code{NULL}, invisibly.
}
\description{
Tests whether all vectors have the same length.
}
