% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prohibit_vector_recycling.R
\name{prohibit_vector_recycling}
\alias{prohibit_vector_recycling}
\title{Prohibit vector recycling}
\source{
\url{http://stackoverflow.com/a/9335687/1664978}
}
\usage{
prohibit_vector_recycling(...)
}
\arguments{
\item{...}{A list of vectors}
}
\value{
An error message if the vectors are of different length (unless the alternative length is 1).
}
\description{
Tests (harshly) whether the vectors can be recycled safely.
}
\examples{
\dontrun{
# Returns nothing because they are of the same length
prohibit_vector_recycling(c(2, 2), c(2, 2))
# Returns nothing also, because the only different length is 1
prohibit_vector_recycling(c(2, 2), 1)
# Returns an error:
prohibit_vector_recycling(c(2, 2), 1, c(3, 3, 3))
}
}
