% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_income.R
\name{rebate_income}
\alias{rebate_income}
\title{Rebate income}
\source{
\url{https://www.ato.gov.au/Individuals/Tax-return/2015/Tax-return/Tax-offset-questions-T1-T2/Rebate-income-2015/}
}
\usage{
rebate_income(
  Taxable_Income,
  Rptbl_Empr_spr_cont_amt = 0,
  All_deductible_super_contr = 0,
  Net_fincl_invstmt_lss_amt = 0,
  Net_rent_amt = 0,
  Rep_frng_ben_amt = 0
)
}
\arguments{
\item{Taxable_Income}{the taxable income}

\item{Rptbl_Empr_spr_cont_amt}{The reportable employer superannuation contributions amount}

\item{All_deductible_super_contr}{deductible personal superannuation contributions}

\item{Net_fincl_invstmt_lss_amt}{Net financial investment loss}

\item{Net_rent_amt}{(for Rental deductions)}

\item{Rep_frng_ben_amt}{Reportable fringe-benefits}
}
\description{
Rebate income
}
