% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_taxstats.R
\name{require_taxstats}
\alias{require_taxstats}
\alias{require_taxstats1516}
\title{Attach a 'taxstats' package}
\usage{
require_taxstats()

require_taxstats1516()
}
\value{
\code{TRUE}, invisibly, for success. Used for its side-effect: attaching the taxstats 
package.
}
\description{
Used in lieu of simply \code{library(taxstats)} to 
handle cases where it is not installed, but should not be installed
to the user's default library (as during CRAN checks).
}
