% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residential_property_prices.R
\docType{data}
\name{residential_property_prices}
\alias{residential_property_prices}
\title{Residential property prices in Australia}
\format{A \code{data.table} of three columns and 522 observations:
\describe{
\item{Date}{Date of the index}
\item{City}{Capital city (or \code{Australia (weighted average)})}
\item{Residential_property_price_index}{An index (100 = 2011-12-01) measuring the price change in all residential dwellings.}
}}
\source{
ABS Cat 6416.0. \url{http://www.abs.gov.au/ausstats/abs@.nsf/mf/6416.0}.
}
\usage{
residential_property_prices
}
\description{
Residential property prices indexes for the capital cities of Australia, and a weighted average for the whole country. Last updated 2018-07-06.
}
\keyword{datasets}
