% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revenue_foregone.R
\name{revenue_foregone}
\alias{revenue_foregone}
\title{Revenue foregone from a modelled sample file}
\usage{
revenue_foregone(dt, revenue_positive = TRUE, digits = NULL)
}
\arguments{
\item{dt}{A \code{data.table} from \code{\link{model_income_tax}}.}

\item{revenue_positive}{If \code{TRUE}, the default, tax increase (revenue) is positive and tax cuts are negative.}

\item{digits}{If not \code{NULL}, affects the print method of the value.}
}
\description{
Revenue foregone from a modelled sample file
}
