% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapto.R
\name{sapto}
\alias{sapto}
\title{Senior Australians and Pensioner Tax Offset}
\usage{
sapto(rebate_income, fy.year, sapto.eligible, family_status = "single")
}
\arguments{
\item{rebate_income}{The rebate income of the individual.}

\item{fy.year}{The financial year in which sapto is to be calculated.}

\item{sapto.eligible}{Is the individual eligible for sapto?}

\item{family_status}{Family status of the individual.}
}
\description{
Senior Australians and Pensioner Tax Offset
}

