% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapto_rcpp.R
\name{sapto_rcpp}
\alias{sapto_rcpp}
\title{SAPTO done in Rcpp}
\usage{
sapto_rcpp(
  RebateIncome,
  MaxOffset,
  LowerThreshold,
  TaperRate,
  SaptoEligible,
  SpouseIncome,
  IsMarried
)
}
\arguments{
\item{RebateIncome, MaxOffset, LowerThreshold, TaperRate, SaptoEligible, SpouseIncome, IsMarried}{Arguments
as in \code{\link{sapto}}.}
}
\description{
SAPTO done in Rcpp
}
