% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sapto_rcpp_singleton}
\alias{sapto_rcpp_singleton}
\title{SAPTO singleton}
\usage{
sapto_rcpp_singleton(
  rebate_income,
  max_offset,
  lower_threshold,
  taper_rate,
  sapto_eligible,
  Spouse_income,
  is_married
)
}
\arguments{
\item{rebate_income, max_offset, lower_threshold, taper_rate, sapto_eligible, Spouse_income, is_married}{As in \code{\link{sapto}}.}
}
\description{
Length-one version of \code{SAPTO} in C++.
}
