% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sapto_rcpp_yr}
\alias{sapto_rcpp_yr}
\title{SAPTO for specific years in C++}
\arguments{
\item{RebateIncome, IsMarried, SpouseIncome}{As in \code{\link{sapto}}.}
}
\description{
Fast way to calculate SAPTO for multiple people when the year is known in advance. Speed is by cheating and entering in the year's parameters literally.
}
