% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_with_right_labels.R
\name{stacked_bar_with_right_labels}
\alias{stacked_bar_with_right_labels}
\title{Stacked charts with labels at right}
\usage{
stacked_bar_with_right_labels(.data, geom = "bar", verbose = FALSE,
  right_margin = 0.5, scale_y_args, scale_x_args, theme.args)
}
\arguments{
\item{.data}{A data frame, containing entries for \code{x}, \code{y}, and \code{fill}. \code{x} and \code{fill} must be ordered factors.}

\item{geom}{The type of chart ("bar", "area").}

\item{verbose}{Report the margin used (in grid:: 'lines').}

\item{right_margin}{The amount of padding at right to use. The whole point of this function is to select a good right margin to allow space. But if the margin provided is wrong, it can be changed manually here.}

\item{scale_y_args}{A list of arguments passed to r \code{ggplot2::scale_y_continuous}.}

\item{scale_x_args}{A list of arguments passed to \code{ggplot2::scale_x_continuous}.}

\item{theme.args}{A list of arguments passed to \code{ggplot2::theme}.}
}
\value{
A chart with the labels in the right gutter
}
\description{
Stacked charts with labels at right
}
\examples{
library(data.table)
dat <- data.table::CJ(
  x = factor(1:10, ordered = TRUE),
  fill = factor(c("A long but not\\ntoo long label", letters[2:3]),
                levels = c("A long but not\\ntoo long label", letters[2:3]),
                ordered = TRUE)
) 
dat$y <- abs(rnorm(1:nrow(dat)))

stacked_bar_with_right_labels(dat)
}

