% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_repayment.R
\name{student_repayment}
\alias{student_repayment}
\title{HELP / HECS repayment amounts}
\source{
\url{https://www.ato.gov.au/Rates/HELP,-TSL-and-SFSS-repayment-thresholds-and-rates/?page=2#HELP_repayment_thresholds_and_rates_2013_14}
}
\usage{
student_repayment(repayment_income, fy.year, debt)
}
\arguments{
\item{repayment_income}{The repayment income of the individual, equal to 
\deqn{{Taxable Income} + {Total net investment loss (incl Net rental loss)} + {reportable fringe benefits amounts} + {Reportable super contributions} + {exempt foreign income}}}

\item{fy.year}{The financial year \code{repayment_income} was earned.}

\item{debt}{The amount of student debt held.}
}
\value{
The repayment amount.
}
\description{
HELP / HECS repayment amounts
}
\author{
Ittima Cherastidtham and Hugh Parsonage
}

