% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Switch}
\alias{Switch}
\title{vectorized switch 
e.g. Switch(c("A", "B", "C", "A"), "A" = 1, "B" = 2, "C" = 11:14)}
\usage{
Switch(Expr, ...)
}
\description{
vectorized switch 
e.g. Switch(c("A", "B", "C", "A"), "A" = 1, "B" = 2, "C" = 11:14)
}
