% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texNum.R
\name{texNum}
\alias{texNum}
\title{Convert number to English prose}
\usage{
texNum(number, sig.figs = 3L, dollar = FALSE, pre.phrase = NULL,
  .suffix = NULL)
}
\arguments{
\item{number}{A single numeric vector.}

\item{sig.figs}{Significant figures to be displayed}

\item{dollar}{Logical, should a LaTeX dollar sign (\code{\\$}) be prefixed.}

\item{pre.phrase}{A length 2 character vector to insert a phrase before the number. The first element is the phrase to be used if the output rounds up before printing (i.e. the original number is smaller than the number printed); the second is the phrase to be used if the number rounds down. The default is \code{NULL}, i.e. no phrase inserted. If either string has a trailing tilde (e.g. \code{"almost~"}), the tilde separates the number and the phrase (as a LaTeX control sequence).}

\item{.suffix}{Either \code{NULL} for an automatic suffix based on the magnitude of the number, or one of million, billion, trillion for a fixed suffix.}
}
\value{
A character string representing the number as appropriate for English prose.
}
\description{
This takes an arbitrary numbers and prints like a human would transcribe it. Designed to be used within \code{\\Sexpr}.
}
\examples{
texNum(500e3)
texNum(500e3 - 1, pre.phrase = c("almost", "over"))
texNum(500e6)
}

