% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_grattan_OrangeBackground.R
\name{theme_grattan_OrangeBackground}
\alias{theme_grattan_OrangeBackground}
\title{A theme for use with ggplots when the background needs to be Orange}
\usage{
theme_grattan_OrangeBackground(base_size = 24, base_family = "")
}
\arguments{
\item{base_size}{Size in points for the size of text.}

\item{base_family}{Typeface.}
}
\description{
A theme for use with ggplots when the background needs to be Orange
}

