% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hugh.R
\name{theme_hugh}
\alias{theme_hugh}
\title{A theme for use with ggplots}
\arguments{
\item{base_size}{Size in points for the size of text.}

\item{base_family}{Typeface.}
}
\description{
A theme for use with ggplots
}

