% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/theme_grattan_OrangeBackground.R, R/theme_hugh.R
\name{theme_huh}
\alias{theme_grattan_OrangeBackground}
\alias{theme_hugh}
\alias{theme_huh}
\title{A theme for use with ggplots when the background needs to be Orange}
\usage{
theme_grattan_OrangeBackground(base_size = 24, base_family = "")

theme_hugh(base_size = 23, base_family = "")
}
\description{
A theme for use with ggplots when the background needs to be Orange

A theme for use with ggplots
}

