% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_per.R
\name{validate_per}
\alias{validate_per}
\title{Validate per}
\usage{
validate_per(per, missing_per, .fortnights_per_yr = NULL)
}
\arguments{
\item{per}{How often are payments will made? Can only take values 'year', 'fortnight', or 'quarter'.}

\item{missing_per}{Is `per`` missing in the outer function? If so the defualt for that function is being used.}

\item{.fortnights_per_yr}{What is the ratio of the fortnightly payment amount to the yearly payment amount. Can only take values 26 or 365/14.}
}
\description{
Checks whether a valid input of `per` is used and outputs the amount yearly payments are divided by to get the desired rate.
}
