% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_date.R
\name{verify_date}
\alias{verify_date}
\title{Verifying validity of dates}
\usage{
verify_date(date_to_verify, from = -Inf, to = Inf)
}
\arguments{
\item{date_to_verify}{(character) A user-provided value, purporting to be
character vector of dates.}

\item{from, to}{(integer) Indicating the range of years valid for date_to_verify.
If \code{NULL}, no restriction on the upper or lower bound of the range.}
}
\description{
Many functions expect Dates. 
Determining that they are validly entered is often quite 
computatationally costly, relative to the core calculations.
These internal functions provide mechanisms to check validity
quickly, while still providing clear, accurate error messages.
}
