% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_fys_permitted.R
\name{verify_fys_permitted}
\alias{verify_fys_permitted}
\title{Verifying validity of financial years}
\usage{
verify_fys_permitted(to_verify, permitted_fys, min.yr = NULL,
  max.yr = NULL, deparsed = deparse(substitute(to_verify)))
}
\arguments{
\item{to_verify}{A user-provided value, purporting to be
character vector of financial years.}

\item{permitted_fys}{A character vector of valid financial years.}

\item{min.yr, max.yr}{Integers specifying the range of \code{to_verify}.
If \code{NULL}, no restriction on the upper or lower bound of the range.}

\item{deparsed}{A string indicating the argument that the user provided.
Should generally be provided explicitly as the default is unlikely 
to be user-friendly.}
}
\value{
\code{NULL} invisibly provided \code{to_verify} 
contains financial years as specified. Otherwise an error
explaining why \code{to_verify} could not be validated.
}
\description{
Many functions expect financial years. 
Determining that they are validly entered is often quite 
computatationally costly, relative to the core calculations.
These internal functions provide mechanisms to check validity
quickly, while still providing clear, accurate error messages.
}
