% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_allowance.R
\name{youth_allowance}
\alias{youth_allowance}
\title{Youth allowance}
\usage{
youth_allowance(ordinary_income, age = 18, eligible_if_over22 = FALSE,
  is_single = TRUE, living_at_home = FALSE, has_children = FALSE,
  isjspceoalfofcoahodeoc = FALSE, per = "fortnight")
}
\arguments{
\item{eligible_if_over22}{To be eligible for Youth Allowance while over 22, recipients must either commence full-time study or an Australian apprenticeship having been in receipt of an income support payment for at least 6 out of the last 9 months since turning 22, or study an approved course in English where English is not their first language.}

\item{isjspceoalfofcoahodeoc}{Is the recipient a single job seeker principal carer, either of large family or foster child/ren, or who is a home or distance educator of child/ren?}
}
\description{
Youth allowance
}
