% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_pension_age.R
\name{age_pension_age}
\alias{age_pension_age}
\title{Age of eligibility for the Age Pension}
\source{
\url{http://guides.dss.gov.au/guide-social-security-law/3/4/1/10}
}
\usage{
age_pension_age(when = Sys.Date(), sex = "male")
}
\arguments{
\item{when}{Either a Date}
}
\description{
Age of eligibility for the Age Pension
}
