% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/aus_pop_qtr.R
\name{aus_pop_qtr}
\alias{aus_pop_qtr}
\title{Australia's population}
\usage{
aus_pop_qtr(date_quarter)
}
\arguments{
\item{date}{as character string(YYYY-QQ)}
}
\value{
the population at \code{date}
}
\description{
Australia's population
}

