% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aus_pop_qtr.R
\name{aus_pop_qtr}
\alias{aus_pop_qtr}
\title{Australia's population}
\usage{
aus_pop_qtr(date_quarter, allow.projections = TRUE)
}
\arguments{
\item{date_quarter}{A character string (YYYY-QQ).}

\item{allow.projections}{If the date is beyond the ABS's confirmed data, should a projection be used?}
}
\value{
The population at \code{date_quarter}, or at the most recent year in the data if projections are disallowed.
}
\description{
Australia's population
}
