% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awote.R
\name{awote}
\alias{awote}
\title{AWOTE}
\usage{
awote(
  Date = NULL,
  fy.year = NULL,
  rollDate = "nearest",
  isMale = NA,
  isAdult = TRUE,
  isOrdinary = TRUE
)
}
\arguments{
\item{Date, fy.year}{When the AWOTE is desired.}

\item{rollDate}{How should the \code{Date} be joined to the source data?
Passed to \code{data.table}.}

\item{isMale}{(logical, default: \code{NA}) \code{TRUE} for male weekly earnings, 
\code{FALSE} for female, \code{NA} for the weekly earnings of both sexes.}

\item{isAdult}{(logical, default: \code{TRUE}) Use adult weekly earnings?}

\item{isOrdinary}{Use ordinary weekly earnings?}
}
\description{
Adult weekly ordinary-time earnings
}
\examples{
awote()  # Current AWOTE


}
