% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carer_payment.R
\name{carer_payment}
\alias{carer_payment}
\title{Carer Payment}
\usage{
carer_payment(dclad_eligible = TRUE, adat_eligible = FALSE,
  n_dependents = 0, high_adat = FALSE, living_at_home = FALSE,
  carer_fortnightly_income = 0, carer_annual_income = 0,
  carer_asset_value = 0, care_receiver_fortnightly_income = 0,
  care_receiver_annual_income = 0, care_receiver_asset_value = 0,
  partner_income = 0, partner_assets = 0, children_income = 0,
  children_assets = 0)
}
\description{
Carer Payment
}
\details{

}
