% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_inflator.R
\name{CG_population_inflator}
\alias{CG_population_inflator}
\title{CG_inflator}
\usage{
CG_population_inflator(x = 1, from_fy, to_fy)
}
\arguments{
\item{x}{To be inflated.}

\item{from_fy, to_fy}{Financial years designating the inflation period.}
}
\value{
An estimate of \code{x} inflated in \code{to_fy}
}
\description{
CG_inflator
}

