% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CG_inflator.R
\name{CG_population_inflator}
\alias{CG_population_inflator}
\alias{CG_inflator}
\title{Forecasting capital gains}
\usage{
CG_population_inflator(
  x = 1,
  from_fy,
  to_fy,
  forecast.series = "mean",
  cg.series
)

CG_inflator(x = 1, from_fy, to_fy, forecast.series = "mean")
}
\arguments{
\item{x}{To be inflated.}

\item{from_fy, to_fy}{Financial years designating the inflation period.}

\item{forecast.series}{One of \code{"mean"}, \code{"lower"}, \code{"upper"}. What estimator to use in forecasts. \code{"lower"} and \code{"upper"} give the lower and upper boundaries of the 95\% prediction interval.}

\item{cg.series}{(Not implemented.)}
}
\value{
For \code{CG_population_inflator}, the number of individuals estimated to incur capital gains in \code{fy_year}. 
For \code{CG_inflator}, an estimate of the nominal value of (total) capital gains in \code{to_fy} relative to the nominal value in \code{from_fy}.
}
\description{
Forecasting capital gains
}
