% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpi_inflator.R
\name{cpi_inflator}
\alias{cpi_inflator}
\title{CPI inflator}
\usage{
cpi_inflator(from_nominal_price = 1, from_fy, to_fy = "2013-14",
  adjustment = "none")
}
\arguments{
\item{from_nominal_price}{(numeric) the price (or vector of prices) to be inflated}

\item{from_fy}{(character) a character vector with each element in the form "2012-13" representing the financial year contemporaneous to the from_nominal_price.}

\item{to_fy}{(character) a character vector with each element in the form "2012-13" representing the financial year that prices are to be inflated.}
}
\value{
the value of from_nominal_price in real (to_fy) dollars.
}
\description{
CPI inflator
}

