% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentially_uprate_wage.R
\name{differentially_uprate_wage}
\alias{differentially_uprate_wage}
\title{Differential uprating}
\usage{
differentially_uprate_wage(wage = 1, from_fy, to_fy)
}
\arguments{
\item{wage}{A numeric vector to be uprated.}

\item{from_fy}{The financial year of the origin of the wage, which must be a financial year of an available sample file. (Not after 2013-14.)}

\item{to_fy}{The target of the uprating. Passed to \code{\link{wage_inflator}}.}
}
\description{
Apply differential uprating to projections of the \code{Sw_amt} variable.
}
\author{
Hugh Parsonage and William Young
}

