% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_tax_benefit.R
\name{family_tax_benefit}
\alias{family_tax_benefit}
\title{FAMILY TAX BENEFIT}
\usage{
family_tax_benefit(data, income_test_1_bound = 51027,
  income_test_2_bound = 94316)
}
\description{
FAMILY TAX BENEFIT
}
