% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpal.R
\name{gpal}
\alias{gpal}
\title{automatic palette adjustment (for small number of categories)}
\usage{
gpal(n, dark = TRUE, reverse = FALSE)
}
\arguments{
\item{n}{the number of variables/factors over which the palette is to paint}

\item{dark}{Should a dark palette be used? (Only available for n=2.)}

\item{reverse}{(logical) Option to reverse the palette.}
}
\value{
A vector of HTML colours to be used.
}
\description{
automatic palette adjustment (for small number of categories)
}
\author{
Hugh Parsonage
}

