% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan-package.R
\docType{package}
\name{grattan-package}
\alias{grattan-package}
\alias{_PACKAGE}
\alias{grattan}
\title{The grattan package.}
\description{
Grattan package
}
\details{
Tax modelling and other common tasks for Australian policy analysts, 
in support of the Grattan Institute, Melbourne. 
<https://grattan.edu.au>
}
\section{Package options}{

\describe{
\item{\code{grattan.verbose}}{Emit diagnostic messages (via \code{cat()}))}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HughParsonage/grattan}
  \item Report bugs at \url{https://github.com/HughParsonage/grattan/issues}
}

}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}

\email{hugh.parsonage@gmail.com}
}
\keyword{package}
