% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan-package.R
\docType{package}
\name{grattan-package}
\alias{grattan-package}
\alias{grattan}
\title{The grattan package.}
\description{
Grattan package
}
\details{
Colours, charts, and other common quantative tasks for Australia tax analysts
}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}
}
\keyword{package}
