% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan-package.R
\docType{package}
\name{grattan-package}
\alias{grattan}
\alias{grattan-package}
\title{Colours, charts, and other common quant tasks}
\description{
Grattan package
}
\details{
Colours, charts, and other common quant tasks
}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}
}
\keyword{package}

