% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan-package.R
\docType{package}
\name{grattan-package}
\alias{grattan-package}
\alias{_PACKAGE}
\alias{grattan}
\title{The grattan package.}
\description{
Grattan package
}
\details{
Colours, charts, and other common quantitative tasks for Australia tax analysts
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/HughParsonage/grattan}
  \item Report bugs at \url{https://github.com/HughParsonage/grattan/issues}
}

}
\author{
\email{hugh.parsonage+grattanpackage@grattan.edu.au}
}
\keyword{package}
