% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grattan_waterfall.R
\name{grattan waterfall}
\alias{grattan waterfall}
\alias{grattan_waterfall}
\title{Create waterfall charts}
\usage{
grattan_waterfall(.data = NULL, values, labels, rect_text_labels = values,
  rect_text_labels_anchor = "centre",
  put_rect_text_outside_when_value_below = 0.05 * (max(cumsum(values)) -
  min(cumsum(values))), calc_total = FALSE, total_axis_text = "Total",
  total_rect_text = sum(values), total_rect_color = gpal(6)[1],
  total_rect_text_color = "white", fill_colours = gpalx(length(values)),
  fill_by_sign = TRUE, dark = FALSE, rect_width = 0.7,
  rect_border = "black", draw_lines = TRUE, lines_anchors = c("centre",
  "centre"), linetype = "dashed", draw_axis.x = "behind",
  ggplot_object_name = "mywaterfall")
}
\arguments{
\item{.data}{a data frame containing two columns, one with the values, the other with the labels}

\item{values}{a numeric vector making up the heights of the rectangles in the waterfall}

\item{labels}{the labels corresponding to each vector, marked on the x-axis}

\item{rect_text_labels}{(character) a character vector of the same length as values that are placed on the rectangles}

\item{rect_text_labels_anchor}{(character) How should rect_text_labels be positioned. In future releases, we might have support for north or south anchors, or for directed positioning (negative down, positive up) etc. For now, only centre is supported.}

\item{put_rect_text_outside_when_value_below}{(numeric) the text labels accompanying a rectangle of this height will be placed outside the box: below if it's negative; above if it's positive.}

\item{calc_total}{(logical) should the final pool of the waterfall be calculated (and placed on the chart)}

\item{total_axis_text}{(character) the text appearing on the axis underneath the total rectangle}

\item{total_rect_text}{(character) the text in the middle of the rectangle of the total rectangle}

\item{total_rect_color}{the color of the final rectangle}

\item{total_rect_text_color}{the color of the final rectangle's label text}

\item{fill_by_sign}{(logical) should positive and negative values each have the same colour?}

\item{dark}{(logical) (only for fill_by_sign) should the dark palette be used?}

\item{rect_width}{(numeric) the width of the rectangle, relative to the space between each label factor}

\item{rect_border}{the border around each rectangle. Choose NA if no border is desired.}

\item{draw_lines}{(logical) should lines be drawn between successive rectangles}

\item{lines_anchors}{a character vector of length two specifying the horizontal placement of the drawn lines relative to the preceding and successive rectangles, respectively}

\item{linetype}{the linetype for the draw_lines}

\item{draw_axis.x}{(character) one of "none", "behind", "front" whether to draw an x.axis line and whether to draw it behind or in front of the rectangles, default is behind}

\item{ggplot_object_name}{(character) a quoted valid object name to which ggplot layers may be addded after the function has run}
}
\description{
Create waterfall charts
}

