% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_dollar.R
\name{grattan_dollar}
\alias{grattan_dollar}
\title{dollar scales}
\usage{
grattan_dollar(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector}

\item{digits}{Minimum number of digits after the decimal point. (\code{nsmall} in \code{base::format}).}
}
\description{
dollar scales
}
\details{
Makes negative numbers appear as \eqn{-\$10,000} instead of \eqn{\$-10,000} in \code{scales::dollar}.
}

