% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grattan_percent.R
\name{grattan_percent}
\alias{grattan_percent}
\title{English prose for small numbers}
\usage{
grattan_percent(number, digits = 1, .percent.suffix = "~per~cent")
}
\arguments{
\item{number}{A length-one numeric vector.}

\item{digits}{Passed to \code{scales::comma}.}

\item{.percent.suffix}{What to use for the suffix of the number.}
}
\value{
For \code{grattan_percent}, the number multiplied by 100, with rounding provided by \code{scales::comma} and then the suffix.
}
\description{
When a small number is requested inline, it is often preferable to use an English expression rather than the numeral.
Further, referential purity can be lost when numbers are transformed to suit printing (say by multiplying by 100 to use 'per cent').
A complement to \code{\link{texNum}}. 

These functions are all of the form \code{grattan_*} to avoid namespace clashes.
}

