% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{tax function}
\usage{
income_tax(income, age = 44, fy.year = "2012-13", return.mode = "numeric")
}
\arguments{
\item{income}{the personal assessable income}

\item{age}{the individual's age}

\item{fy.year}{the financial year in which the income was earned}

\item{return.mode}{use numeric or integer}
}
\value{
the total personal income tax payable
}
\description{
tax function
}
\author{
Various
}

