% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{tax function}
\usage{
income_tax(income, fy.year, age = 42, family_status = "individual",
  sample_file, .dots.ATO = NULL)
}
\arguments{
\item{income}{the personal assessable income}

\item{fy.year}{the financial year in which the income was earned}

\item{age}{the individual's age}

\item{sample_file}{(Not yet used) A sample file \code{data.table} for which the income tax payable is desired on each row.}

\item{.dots.ATO}{A data.frame that contains additional information about the individual's circumstances, with columns the same as in the ATO sample files.}

\item{brackets}{a numeric vector designating the lower boundaries of tax brackets}

\item{return.mode}{use numeric or integer}

\item{allow.forecasts}{should dates beyond 2014-15 be permitted?}
}
\value{
the total personal income tax payable
}
\description{
tax function
}
\details{
The function rolling is inflexible by design. It is designed to guarantee the correct tax payable in a year.
}
\author{
Tim Cameron, Brendan Coates, Hugh Parsonage
}

