% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{tax function}
\usage{
income_tax(income, fy.year = "2012-13",
  include.temp.budget.repair.levy = FALSE, return.mode = "numeric",
  age = 44, age_group, is.single = TRUE)
}
\arguments{
\item{income}{the personal assessable income}

\item{fy.year}{the financial year in which the income was earned}

\item{return.mode}{use numeric or integer}

\item{age}{the individual's age}
}
\value{
the total personal income tax payable
}
\description{
tax function
}
\author{
Various
}

