% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/income_tax.R
\name{income_tax}
\alias{income_tax}
\title{tax function}
\usage{
income_tax(income, fy.year = "2012-13")
}
\arguments{
\item{income}{the personal assessable income}

\item{ML.lower}{the lower medicare threshold}

\item{ML.upper}{the upper medicare threshold}
}
\value{
the total personal income tax payable
}
\description{
tax function
}
\author{
Various
}

