% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fy.year.R
\name{is.fy}
\alias{is.fy}
\title{Convenience function for creating FY}
\usage{
is.fy(fy.yr)
}
\arguments{
\item{yr_ending}{()}
}
\value{
the financial year ("YYYY-YY") corresponding to June of \code{yr_ending}
}
\description{
Convenience function for creating FY
}

