% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_inflator.R
\name{lf_inflator}
\alias{lf_inflator}
\alias{lf_inflator_fy}
\title{Labour force inflators}
\source{
ABS Cat 6202.0 \url{http://www.abs.gov.au/ausstats/abs@.nsf/mf/6202.0?OpenDocument}.
}
\usage{
lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy,
  useABSConnection = FALSE, allow.projection = TRUE, use.month = 1L,
  forecast.series = c("mean", "upper", "lower", "custom"),
  forecast.level = 95, lf.series = NULL)

lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date,
  useABSConnection = FALSE)
}
\arguments{
\item{labour_force}{A numeric vector.}

\item{from_fy}{Financial year of \code{labour_force}.}

\item{to_fy}{Financial year for which the labour force is predicted.}

\item{useABSConnection}{Should an sdmx connection be used to get ABS data?}

\item{allow.projection}{Logical. Should projections be allowed?}

\item{use.month}{An integer (corresponding to the output of \code{data.table::month}) representing the month of the series used for the inflation.}

\item{forecast.series}{Whether to use the forecast mean, or the upper or lower boundaries of the prediction intervals.}

\item{forecast.level}{The prediction interval to be used if \code{forecast.series} is \code{upper} or \code{lower}.}

\item{from_date}{The date of \code{labour_force}.}

\item{to_date}{Dates as a character vector.}

\item{lf.indices}{Custom.}
}
\value{
The relative labour force between \code{to_date} and \code{for_date} or \code{to_fy} and \code{from_fy}, multiplied by \code{labour_force}.
}
\description{
Labour force inflators
}
\details{
\code{lf_inflator} is used on dates. The underlying data series is available every month.
}
\examples{
lf_inflator_fy(labour_force = 1, from_fy = "2012-13", to_fy = "2013-14")
\dontrun{
lf_inflator(labour_force = 1, from_date = "2013-06-30", to_date = "2014-06-30")
}
}
\author{
Hugh Parsonage and Tim Cameron
}

