% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_inflator.R
\name{lf_inflator}
\alias{lf_inflator}
\title{Labour force inflators}
\usage{
lf_inflator(labour.force = 1, from_date = "2013-06-30", to_date)
}
\arguments{
\item{labour.force}{a numeric vector}

\item{from_date, }{to_date dates as a character vector}
}
\value{
the relative labour force between to_date and for_date, multiplied by labour.force.
}
\description{
Labour force inflators
}

