% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lf_inflator.R
\docType{import}
\name{lf_inflator}
\alias{lf_inflator}
\alias{lf_inflator_fy}
\title{Objects exported from other packages}
\usage{
lf_inflator_fy(...)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grattanInflators}{\code{\link[grattanInflators]{lf_inflator}}}
}}

