% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare_levy.R
\name{medicare_levy}
\alias{medicare_levy}
\title{Medicare levy}
\usage{
medicare_levy(income, fy.year = "2013-14", Spouse_income = 0,
  sapto.eligible = FALSE, family_status = "individual", n_dependants = 0)
}
\arguments{
\item{income}{The taxable income. A vector of numeric values.}

\item{fy.year}{The financial year. A character vector satisfying \code{is.fy}.}

\item{Spouse_income}{The spouse's adjusted income.}

\item{sapto.eligible}{(logical) Is the taxpayer eligible for SAPTO?}

\item{family_status}{What is the taxpayer's family status: family or individual?}

\item{n_dependants}{Number of children dependant on the taxpayer.}
}
\value{
The medicare levy payable for that taxpayer.
}
\description{
The (actual) amount payable for the medicare levy.
}
\details{
The actual medicare levy payable for such a taxpayer in that financial year.
}

