% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_child_care_subsidy.R
\name{model_child_care_subsidy}
\alias{model_child_care_subsidy}
\title{Model Child Care Subsidy}
\usage{
model_child_care_subsidy(sample_file = 1, cbdc_hourly_cap = 11.77,
  fdc_hourly_cap = 10.9, oshc_hourly_cap = 10.29, ihc_hourly_cap = 25.48,
  annual_cap_income = 186958, annual_cap_subsidy = 10190,
  income_test_bracket_1 = 66958, income_test_bracket_2 = 171958,
  income_test_bracket_3 = 251248, income_test_bracket_4 = 341248,
  income_test_bracket_5 = 354248, taper_1 = 0.85, taper_2 = 0.5,
  taper_3 = 0.2, activity_test_1_brackets = c(0, 8, 16.00001, 48.00001),
  activity_test_1_hours = c(0, 36, 72, 100), calc_baseline_subsidy = FALSE,
  return. = c("sample_file", "new_subsidy"))
}
\description{
The child care subsidy if threshholds and rates are changed.
}
