% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_income_tax.R
\name{model_income_tax}
\alias{model_income_tax}
\title{Modelled Income Tax}
\usage{
model_income_tax(sample_file, baseline_fy, exclude = c("nothing",
  "ordinary_tax", "medicare_levy", "lito", "sapto"), n_dependants = 0L,
  ordinary_tax_thresholds = NULL, ordinary_tax_rates = NULL,
  medicare_levy_lower_threshold = NULL,
  medicare_levy_upper_threshold = NULL, medicare_levy_taper = NULL,
  medicare_levy_rate = NULL, medicare_levy_lower_family_threshold = NULL,
  medicare_levy_upper_family_threshold = NULL,
  medicare_levy_lower_up_for_each_child = NULL, lito_max_offset = NULL,
  lito_taper = NULL, lito_min_bracket = NULL, sapto_eligible = NULL,
  sapto_max_offset = NULL, sapto_lower_threshold = NULL,
  sapto_taper = NULL)
}
\arguments{
\item{sample_file}{A sample file having at least as many variables as the 2012-13 sample file.}

\item{baseline_fy}{If a parameter is not selected, the parameter's value in this tax year is used.}

\item{exclude}{A character vector specifying which womponents of the income tax to \emph{exclude}. 
Multiple values are allowed. 
The special value \code{"nothing"} means no components are excluded. 
If present at the first position of \code{exclude} (the default) it takes precedence and no components are excluded; 
elsewhere it has no effect.}

\item{n_dependants}{The number of dependants for each entry in \code{sample_file}.}

\item{ordinary_tax_thresholds}{A numeric vector specifying the lower bounds of the brackets for "ordinary tax" as defined by the Regulations.
The first element should be zero if there is a tax-free threshold.}

\item{ordinary_tax_rates}{The marginal rates of ordinary tax. The first element should be zero if there is a tax-free threshold. 
Since the temporary budget repair levy was imposed on a discrete tax bracket when it applied, it is not included in this function.}

\item{medicare_levy_lower_threshold}{Minimum taxable income at which the Medicare levy will be applied.}

\item{medicare_levy_upper_threshold}{Minimum taxable income at which the Medicare levy will be applied at the full Medicare levy rate (2\% in 2015-16). Between this threshold and the \code{medicare_levy_lower_threshold}, a tapered rate applies, starting from zero and climbing to \code{medicare_levy_rate}.}

\item{medicare_levy_taper}{The taper that applies between the \code{_lower} and \code{_upper} thresholds.}

\item{medicare_levy_rate}{The ordinary rate of the Medicare levy for taxable incomes above \code{medicare_levy_upper_threshold}.}

\item{medicare_levy_lower_family_threshold, medicare_levy_upper_family_threshold}{The equivalent values for families.}

\item{medicare_levy_lower_up_for_each_child}{The amount to add to the \code{_family_threshold}s for each dependant child.}

\item{lito_max_offset}{The maximum offset available for low incomes.}

\item{lito_taper}{The taper to apply beyond \code{lito_min_bracket}.}

\item{lito_min_bracket}{The taxable income at which the value of the offset starts to reduce (from \code{lito_max_offset}).}

\item{sapto_eligible}{Whether or not each taxpayer in \code{sample_file} is eligible for \code{SAPTO}. 
If \code{NULL}, the default, then eligibility is determined by \code{age_range} in \code{sample_file};
\emph{i.e.}, if \code{age_range <= 1} then the taxpayer is assumed to be eligible for SAPTO.}

\item{sapto_max_offset}{The maximum offset available through SAPTO.}

\item{sapto_lower_threshold}{The threshold at which SAPTO begins to reduce (from \code{sapto_max_offset}).}

\item{sapto_taper}{The taper rate beyond \code{sapto_lower_threshold}.}
}
\description{
The income tax payable if tax settings are changed.
}
