% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_income_tax.R
\name{new_income_tax}
\alias{new_income_tax}
\title{New income tax payable
Income tax payable with new tax brackets, tax rates etc}
\usage{
new_income_tax(income, new_tax_tbl)
}
\arguments{
\item{income}{A vector of taxable incomes.}

\item{new_tax_tbl}{A \code{data.table} with columns \code{lower_bracket} and \code{marginal_rate} for the new brackets and marginal rates.}
}
\value{
The income according to the new parameters.
}
\description{
New income tax payable
Income tax payable with new tax brackets, tax rates etc
}
