% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_sapto.R
\name{new_sapto}
\alias{new_sapto}
\title{SAPTO with user-defined thresholds}
\usage{
new_sapto(
  rebate_income,
  new_sapto_tbl,
  sapto.eligible = TRUE,
  Spouse_income = 0,
  fill = 0,
  family_status = "single"
)
}
\arguments{
\item{rebate_income}{The rebate income of the individual.}

\item{new_sapto_tbl}{Having the same columns as \code{grattan:::sapto_tbl}, keyed on family_status.}

\item{sapto.eligible}{Is the individual eligible for sapto?}

\item{Spouse_income}{Spouse income whose unutilized SAPTO may be added to the current taxpayer. Must match \code{family_status}; i.e. can only be nonzero when \code{family_status != "single"}.}

\item{fill}{If SAPTO was not applicable, what value should be used?}

\item{family_status}{Family status of the individual.}
}
\description{
SAPTO with user-defined thresholds
}
