% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstart_allowance.R
\name{newstart_allowance}
\alias{newstart_allowance}
\title{Newstart allowance}
\source{
\url{http://classic.austlii.edu.au/au/legis/cth/consol_act/ssa1991186/s1068.html}
}
\usage{
newstart_allowance(ordinary_income = 0, has_partner = FALSE,
  n_dependants = 0, nine_months = FALSE, isjspceoalfofcoahodeoc = FALSE,
  principal_carer = FALSE, partner_income = 0, partner_eligible = FALSE,
  age = 22, fy.year = "2015-16", assets_value = 0, homeowner = FALSE,
  lower = 102, upper = 252, taper1 = 0.5, taper2 = 0.6, taper3 = 0.4,
  per = "fortnight")
}
\arguments{
\item{ordinary_income}{'Ordinary income' within the meaning of s. 1068-G1 of the \emph{Social Security Act 1991}.}

\item{principal_carer}{parent wiht most of the day to day care of child. Defined in https://www.humanservices.gov.au/individuals/enablers/principal-carer-rules-parenting-payment/41456}

\item{partner_income}{The partner's income.}

\item{age}{The individual's age. Different rates apply to those aged 18 to 20.}

\item{per}{Fortnight, the default, specified the units the inputs and outputs are returned.}
}
\description{
Newstart allowance
}
