% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newstart_allowance.R
\name{newstart_allowance}
\alias{newstart_allowance}
\title{Newstart allowance}
\source{
\url{http://classic.austlii.edu.au/au/legis/cth/consol_act/ssa1991186/s1068.html}
}
\usage{
newstart_allowance(ordinary_income = 0, has_partner = FALSE,
  n_dependants = 0, nine_months = FALSE, isjspceoalfofcoahodeoc = FALSE,
  principal_carer = FALSE, partner_income = 0, age = 22, fy.year = NULL,
  assets_value = 0, homeowner = FALSE, lower = 102, upper = 252,
  taper1 = 0.5, taper2 = 0.6, taper3 = 0.4, per = "fortnight")
}
\arguments{
\item{ordinary_income}{'Ordinary income' within the meaning of s. 1068-G1 of the \emph{Social Security Act 1991}.}

\item{has_partner}{Does the individual have a partner?}

\item{n_dependants}{How many dependant children does the individual have?}

\item{nine_months}{If the person is over 60 years old, have they been receiving payments for over 9 continuous months?}

\item{isjspceoalfofcoahodeoc}{Is the recipient a single job seeker principal carer, either of large family or foster child/ren, or who is a home or distance educator of child/ren?}

\item{principal_carer}{Is the individual the parent with most of the day to day care of child. Defined in https://www.humanservices.gov.au/individuals/enablers/principal-carer-rules-parenting-payment/41456}

\item{partner_income}{The partner's income.}

\item{age}{The individual's age.}

\item{fy.year}{Financial year. Default is "2015-16".}

\item{assets_value}{Total value of household assets. Details can be found at https://www.humanservices.gov.au/individuals/enablers/assets/30621}

\item{homeowner}{Is the individual a homeowner?}

\item{lower}{Lower bound for which reduction in payment occurs at rate taper1 (taper 3 for principal carers).}

\item{upper}{Upper bound for which reduction in payment occurs at rate taper1. Lower bound for which reduction in payment occurs at rate taper2. Note that for principal carers there is no upper bound upper.}

\item{taper1}{The amount at which the payment is reduced for each dollar earned between the lower and upper bounds for non-principal carers.}

\item{taper2}{The amount at which the payment is reduced for each dollar earned above the upper bound for non-principal carers.}

\item{taper3}{The amount at which the payment is reduced for each dollar earned above the lower bound for principal carers.}

\item{per}{Fortnight, the default, specified the units the inputs and outputs are returned.}
}
\description{
Newstart allowance
}
