% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pension_supplement.R
\name{pension_supplement}
\alias{pension_supplement}
\title{Pension Supplement}
\usage{
pension_supplement(has_partner = FALSE, age = 70,
  parenting_payment = FALSE, Date = NULL, fy.year = NULL,
  age_pension_age_requirement = FALSE, per = "fortnight",
  overseas_absence = FALSE)
}
\description{
Gets added to max rate of payment before income reduction tests are applied
Supplementary payment for those receiving Age Pension, Carer Payment, Wife Pension, Widow B Pension, Bereavement Allowance, Disability Support Pension (except if under 21 and no kids)
Can also receive if over age pension age and  receiving ABSTUDY, Austudy, Parenting Payment, Partner Allowance, Special Benefit, or, Widow Allowance
Can still receive the basic amount if single, under age pension age, and receive the Parenting Payment
}
