% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_ntile.R, R/weighted_ntile.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate_ntile}
\alias{weighted_ntile}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hutils}{\code{\link[hutils]{mutate_ntile}}, \code{\link[hutils]{weighted_ntile}}}
}}

