% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_ntile.R, R/pmaxC.R, R/weighted_ntile.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{mutate_ntile}
\alias{pmaxC}
\alias{pmax3}
\alias{pminC}
\alias{pmaxV}
\alias{pminV}
\alias{weighted_ntile}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hutils}{\code{\link[hutils]{mutate_ntile}}, \code{\link[hutils]{weighted_ntile}}}

  \item{hutilscpp}{\code{\link[hutilscpp:pmaxC]{pmax3}}, \code{\link[hutilscpp]{pmaxC}}, \code{\link[hutilscpp:pmaxC]{pmaxV}}, \code{\link[hutilscpp:pmaxC]{pminC}}, \code{\link[hutilscpp:pmaxC]{pminV}}}
}}

