% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpi_inflator.R, R/lf_inflator.R,
%   R/mutate_ntile.R, R/pmaxC.R, R/wage_inflator.R, R/weighted_ntile.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cpi_inflator}
\alias{lf_inflator}
\alias{mutate_ntile}
\alias{pmaxC}
\alias{pmax3}
\alias{pminC}
\alias{pmaxV}
\alias{pminV}
\alias{wage_inflator}
\alias{weighted_ntile}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{grattanInflators}{\code{\link[grattanInflators]{cpi_inflator}}, \code{\link[grattanInflators]{lf_inflator}}, \code{\link[grattanInflators]{wage_inflator}}}

  \item{hutils}{\code{\link[hutils]{mutate_ntile}}, \code{\link[hutils]{weighted_ntile}}}

  \item{hutilscpp}{\code{\link[hutilscpp:pmaxC]{pmax3}}, \code{\link[hutilscpp]{pmaxC}}, \code{\link[hutilscpp:pmaxC]{pmaxV}}, \code{\link[hutilscpp:pmaxC]{pminC}}, \code{\link[hutilscpp:pmaxC]{pminV}}}
}}

