% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapto.R
\name{sapto}
\alias{sapto}
\title{Seniors and Pensioner Tax Offset}
\usage{
sapto(rebate_income, fy.year, fill = 0, sapto.eligible = TRUE,
  family_status = "single")
}
\arguments{
\item{rebate_income}{The rebate income of the individual.}

\item{fy.year}{The financial year in which sapto is to be calculated.}

\item{fill}{If SAPTO was not applicable, what value should be used?}

\item{sapto.eligible}{Is the individual eligible for sapto?}

\item{family_status}{Family status of the individual.}
}
\description{
Seniors and Pensioner Tax Offset
}

