% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sapto_rcpp}
\alias{sapto_rcpp}
\title{SAPTO in C++}
\usage{
sapto_rcpp(RebateIncome, MaxOffset, LowerThreshold, TaperRate, SaptoEligible,
  SpouseIncome, FamilyStatus)
}
\arguments{
\item{RebateIncome, MaxOffset, LowerThreshold, TaperRate, SaptoEligible, SpouseIncome, FamilyStatus}{As in \code{\link{sapto}}.}
}
