% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_discrete_grattan.R
\name{scale_color_discrete_grattan}
\alias{scale_color_discrete_grattan}
\title{scale color scale for grattan (uses the palette)}
\usage{
scale_color_discrete_grattan(..., reverse = FALSE)
}
\arguments{
\item{...}{arguments passed to discrete_scale.}

\item{reverse}{(logical) Option to reverse the palette.}
}
\description{
scale color scale for grattan (uses the palette)
}
\details{
A function controlling the colors.
}

