% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale_color_discrete_grattan.R
\name{scale_color_discrete_grattan}
\alias{scale_color_discrete_grattan}
\title{}
\usage{
scale_color_discrete_grattan(...)
}
\arguments{
\item{a}{function controlling the colors}
}
\description{

}

