% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/small_business_tax_offset.R
\name{small_business_tax_offset}
\alias{small_business_tax_offset}
\title{Small Business Tax Offset}
\source{
Basic income tax method s4-10(3) \url{http://classic.austlii.edu.au/au/legis/cth/consol_act/itaa1997240/s4.10.html}.
Explanatory memorandum \url{http://parlinfo.aph.gov.au/parlInfo/download/legislation/ems/r5494_ems_0a26ca86-9c3f-4ffa-9b81-219ac09be454/upload_pdf/503041.pdf}.
}
\usage{
small_business_tax_offset(taxable_income, basic_income_tax_liability,
  total_net_small_business_income, fy_year = NULL, tax_discount = NULL)
}
\arguments{
\item{taxable_income}{Individual's assessable income.}

\item{basic_income_tax_liability}{Tax liability (in dollars) according to the method in the box in s 4.10(3) of the \emph{Income Tax Assessment Act 1997} (Cth). In general, \code{basic_income_tax_liability} is the ordinary tax minus offsets. In particular, it does not include levies (such as the Medicare levy or the Temporary Budget Repair Levy). 
\if{latex}{\deqn{\mathrm{Income\;Tax} = \mathrm{Taxable\;income\times\mathrm{Rate}-\mathrm{Tax\;offsets}}}}
For example, in 2015-16, an individual with an assessable income of \$100,000 had a basic tax liability of 
approximately \$25,000.}

\item{fy_year}{The financial year for which the small business tax offset is to apply.}

\item{total_net_business_income}{Total net business income within the meaning of the Act. For most taxpayers, this is simply any net income from a business they own (or their share of net income from a business in which they have an interest). The only difference being in the calculation of the net business income of some minors (vide Division 6AA of Part III of the Act).}
}
\description{
Small Business Tax Offset
}
