% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/texNum.R
\name{texNum}
\alias{texNum}
\title{Convert number to English prose}
\usage{
texNum(number, sig.figs = 3L, dollar = FALSE)
}
\arguments{
\item{sig.figs}{(not used)}
}
\description{
Convert number to English prose
}

