% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_date.R
\name{verify_date}
\alias{verify_date}
\title{Verifying validity of dates}
\usage{
verify_date(date_to_verify, from = NULL, to = NULL)
}
\arguments{
\item{date_to_verify}{(character) A user-provided value, purporting to be
character vector of dates.}

\item{from, to}{Indicating the range of years valid for \code{date_to_verify}. Default set to -Inf and Inf respectively (i.e. there is no bound)}
}
\description{
Many functions expect Dates. 
Determining that they are validly entered is often quite 
computatationally costly, relative to the core calculations.
These internal functions provide mechanisms to check validity
quickly, while still providing clear, accurate error messages.
}
