% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wage_inflator.R
\name{wage_inflator}
\alias{wage_inflator}
\title{a wage inflator}
\usage{
wage_inflator(wage = 1, from_fy, to_fy, useABSConnection = FALSE,
  allow.projection = TRUE)
}
\arguments{
\item{wage}{the amount to be inflated (1 by default)}

\item{from_fy}{a character vector of the form "2012-13" representing the FY ending that the wage index is to be taken (i.e. Q4 in that year). FY year must be 1996-97 or later.}

\item{to_fy}{the FY ending that the wage index is to be taken}
}
\value{
the wage inflation between the two years
}
\description{
a wage inflator
}

