% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youth_allowance.R
\name{youth_allowance}
\alias{youth_allowance}
\title{Youth allowance}
\usage{
youth_allowance(ordinary_income = 0, fy.year = yr2fy(year(Sys.Date())),
  age = 18, eligible_if_over22 = FALSE, is_single = TRUE,
  living_at_home = FALSE, has_children = FALSE,
  isjspceoalfofcoahodeoc = FALSE, is_student = TRUE, per = "fortnight",
  taper1 = 0.5, taper2 = 0.6, FT_YA_student_lower = 437,
  FT_YA_student_upper = 524, FT_YA_jobseeker_lower = 143,
  FT_YA_jobseeker_upper = 250, excess_partner_income_mu = 0.6)
}
\arguments{
\item{eligible_if_over22}{To be eligible for Youth Allowance while over 22, recipients must either commence full-time study or an Australian apprenticeship having been in receipt of an income support payment for at least 6 out of the last 9 months since turning 22, or study an approved course in English where English is not their first language.}

\item{isjspceoalfofcoahodeoc}{Is the recipient a single job seeker principal carer, either of large family or foster child/ren, or who is a home or distance educator of child/ren?}
}
\description{
Youth allowance
}
